requirements.txt
fastapi
transformers[pythorch]
python-keycloakanyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
deprecation==2.1.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
orjson==3.8.10
packaging==23.1
pyasn1==0.5.0
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==2.15.3
python-multipart==0.0.6
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
